cleanup BE
